community/sample-apps/python-backend-server/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.6.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
backports.cached-property==1.0.2
black==22.6.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
dnspython==2.2.1
email-validator==1.2.1
fastapi==0.79.1
frozenlist==1.3.1
graphql-core==3.2.1
h11==0.13.0
httptools==0.4.0
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
multidict==6.0.2
mypy-extensions==0.4.3
orjson==3.7.12
pathspec==0.9.0
platformdirs==2.5.2
pydantic==1.9.2
Pygments==2.13.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
PyYAML==6.0
qlient-aiohttp==0.1.0b2
qlient-core==1.0.1
requests==2.28.1
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
strawberry-graphql==0.126.0
tomli==2.0.1
typing_extensions==4.3.0
ujson==5.4.0
urllib3==1.26.11
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websocket-client==1.3.3
websockets==10.3
yarl==1.8.1
zipp==3.8.1
